+++ resources/js/Components/Form/Select.vue
@@ -8,7 +8,7 @@
             :class="{ 'is-invalid': $attrs.invalid }"
             v-model="value"
         >
-            <option :value="null" disabled>--- {{ $attrs.label }} ---</option>
+            <option :value="null" disabled>{{ $attrs.label }}</option>
             <option v-for="(value, label) in options" :key="label" :value="value">
                 {{ label }}
             </option>