+++ src/Contracts/Models/Cart.php
@@ -4,9 +4,16 @@ namespace Bazar\Contracts\Models;
 
 use Bazar\Contracts\Discountable;
 use Bazar\Contracts\Itemable;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 interface Cart extends Discountable, Itemable
 {
+    /**
+     * Get the order for the cart.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
+    public function order(): BelongsTo;
     /**
      * Lock the cart.
      *