+++ src/Gateway/Driver.php
@@ -7,29 +7,12 @@ use Bazar\Events\CheckoutProcessing;
 use Bazar\Models\Cart;
 use Bazar\Models\Order;
 use Bazar\Models\Transaction;
+use Bazar\Support\Driver as BaseDriver;
 use Illuminate\Http\Request;
 use Throwable;
 
-abstract class Driver
+abstract class Driver extends BaseDriver
 {
-    /**
-     * The driver config.
-     *
-     * @var array
-     */
-    protected array $config = [];
-
-    /**
-     * Create a new driver instance.
-     *
-     * @param  array  $config
-     * @return void
-     */
-    public function __construct(array $config = [])
-    {
-        $this->config = $config;
-    }
-
     /**
      * Process the payment.
      *
@@ -49,18 +32,6 @@ abstract class Driver
      * @return \Bazar\Models\Transaction
      */
     abstract public function refund(Order $order, ?float $amount = null, array $attributes = []): Transaction;
-
-    /**
-     * Get the name of the driver.
-     *
-     * @return string
-     */
-    public function getName(): string
-    {
-        return preg_replace(
-            '/([a-z0-9])([A-Z])/', '$1 $2', str_replace('Driver', '', class_basename(static::class))
-        );
-    }
 
     /**
      * Get the URL of the transaction.