+++ src/Gateway/Manager.php
@@ -3,10 +3,26 @@
 namespace Bazar\Gateway;
 
 use Bazar\Contracts\Gateway\Manager as Contract;
+use Bazar\Contracts\Itemable;
+use Illuminate\Contracts\Container\Container;
 use Illuminate\Support\Manager as BaseManager;
 
 class Manager extends BaseManager implements Contract
 {
+    /**
+     * Create a new manager instance.
+     *
+     * @param  \Illuminate\Contracts\Container\Container  $container
+     * @return void
+     */
+    public function __construct(Container $container)
+    {
+        parent::__construct($container);
+
+        $this->drivers['cash'] = $this->createDriver('cash');
+        $this->drivers['transfer'] = $this->createDriver('transfer');
+    }
+
     /**
      * Get the default driver name.
      *
@@ -18,27 +34,22 @@ class Manager extends BaseManager implements Contract
     }
 
     /**
-     * Get all of the created "drivers".
+     * Get the available drivers for the given model.
      *
-     * @return array
+     * @param  \Bazar\Contracts\Itemable  $model
+     * @return void
      */
-    public function getDrivers(): array
+    public function getAvailableDriversFor(Itemable $model): array
     {
-        if (! isset($this->drivers['cash'])) {
-            $this->drivers['cash'] = $this->createDriver('cash');
-        }
-
-        if (! isset($this->drivers['transfer'])) {
-            $this->drivers['transfer'] = $this->createDriver('transfer');
-        }
-
-        foreach (array_keys(array_diff_key($this->customCreators, $this->drivers)) as $key) {
+        foreach (array_keys(array_diff_key($this->customCreators, parent::getDrivers())) as $key) {
             if (! isset($this->drivers[$key])) {
                 $this->drivers[$key] = $this->createDriver($key);
             }
         }
 
-        return parent::getDrivers();
+        return array_filter(parent::getDrivers(), static function (Driver $driver) use ($model): bool {
+            return $driver->availableFor($model);
+        });
     }
 
     /**